refactor(plugin-exercise-group): own component for editor Intermediate Task
