+++ packages/editor/src/plugins/exercise-group/editor/editor.tsx
@@ -1,15 +1,12 @@
 import { EditorTooltip } from '@editor/editor-ui/editor-tooltip'
 import { SerloAddButton } from '@editor/plugin/helpers/serlo-editor-button'
 import { EditorPluginType } from '@editor/types/editor-plugin-type'
-import {
-  faArrowCircleDown,
-  faArrowCircleUp,
-  faTrashAlt,
-} from '@fortawesome/free-solid-svg-icons'
+import { faArrowCircleUp, faTrashAlt } from '@fortawesome/free-solid-svg-icons'
 import { useEditorStrings } from '@serlo/frontend/src/contexts/logged-in-data-context'
 
-import type { ExerciseGroupProps } from '.'
-import { ExerciseGroupRenderer } from './renderer'
+import { IntermediateTask } from './intermediate-task'
+import { type ExerciseGroupProps } from '..'
+import { ExerciseGroupRenderer } from '../renderer'
 import { FaIcon } from '@/components/fa-icon'
 import { shouldUseFeature } from '@/components/user/profile-experimental'
 
@@ -53,7 +50,11 @@ export function ExeriseGroupEditor({ state }: ExerciseGroupProps) {
                   </button>
                 </nav>
                 {exercise.render()}
-                {renderIntermediateTask(index)}
+                <IntermediateTask
+                  intermediateTasks={intermediateTasks}
+                  exerciseIndex={index}
+                  lastExerciseIndex={lastExerciseIndex}
+                />
               </>
             ),
           }
@@ -97,58 +98,6 @@ export function ExeriseGroupEditor({ state }: ExerciseGroupProps) {
             className="mb-8 mt-4"
           />
         ) : null}
-      </>
-    )
-  }
-
-  function renderIntermediateTask(exerciseIndex: number) {
-    if (!intermediateTasks.defined) return
-    const taskIndex = intermediateTasks.findIndex(
-      (task) => task.afterIndex.value === exerciseIndex
-    )
-    const task = intermediateTasks[taskIndex]
-    if (!task) return null
-
-    return (
-      <>
-        <nav className="flex justify-end">
-          <small className="mx-2 bg-editor-primary-50 p-1 font-bold text-gray-600">
-            {exGroupStrings.intermediateTask}
-          </small>
-          <div>
-            {exerciseIndex > 0 ? (
-              <button
-                className="serlo-button-editor-secondary serlo-tooltip-trigger mr-2"
-                onClick={() => {
-                  task.afterIndex.set(exerciseIndex - 1)
-                }}
-              >
-                <EditorTooltip text={templateStrings.article.moveUpLabel} />
-                <FaIcon icon={faArrowCircleUp} />
-              </button>
-            ) : null}
-            {exerciseIndex < lastExerciseIndex ? (
-              <button
-                className="serlo-button-editor-secondary serlo-tooltip-trigger mr-2"
-                onClick={() => {
-                  task.afterIndex.set(exerciseIndex + 1)
-                }}
-              >
-                <FaIcon icon={faArrowCircleDown} />
-              </button>
-            ) : null}
-            <button
-              className="serlo-button-editor-secondary serlo-tooltip-trigger mr-2"
-              onClick={() => intermediateTasks.remove(taskIndex)}
-            >
-              <EditorTooltip text={exGroupStrings.removeIntermediateTask} />
-              <FaIcon icon={faTrashAlt} />
-            </button>
-          </div>
-        </nav>
-        <div className="rounded-lg bg-gray-50 p-0.25">
-          {task.content.render()}
-        </div>
       </>
     )
   }
